fishTile method
Added a new interface "DisplayFishingResults" 
Added a method in "DisplayToPlayer" which gets called in fishTile and displays the amount of fished caught and how many hours it took.

+++ worldofzuul/DisplayInterfaces/DisplayFishingResult.java
@@ -2,5 +2,5 @@ package worldofzuul.DisplayInterfaces;
 
 public interface DisplayFishingResult {
 
-    void displayFishingResult (int catchRate, int hoursFished);
+    void displayFishingResult (double catchRate, int hoursFished);
 }
